+++ test-data/p1-preloaded_lists/employers.sql
@@ -1,8 +1,0 @@
-USE linkServicesGroup;
-
-INSERT INTO Employer VALUES
-(1, '3pw0k0L', '2018-11-19 07:58:04', 0, 'Connelly-Goldner', '1 Eastlawn Pass', 'Karl Bontoft', 'Payment Adjustment Coordinator', 'kbontoft0@jimdo.com', '0496978096', 'www.kbontoft0.com.au', 1),
-(2, 'bK1KGdN45gD', '2020-12-27 12:16:49', 0, 'Green-Brekke', '87 Charing Cross Hill', 'Maryanne Foxten', 'Dental Hygienist', 'mfoxten1@hud.gov', '0447477525', 'www.mfoxten1.com.au', 1),
-(3, '4nVwVp4', '2019-01-29 00:39:02', 1, 'Botsford, Bergstrom and Ortiz', '9 Portage Court', 'Charlot Ridwood', 'Food Chemist', 'cridwood2@alibaba.com', '0420679615', 'www.cridwood2.com.au', 1),
-(4, 'uL6Wc6p', '2020-10-06 08:00:28', 0, 'Gerhold-Ankunding', '020 Stone Corner Street', 'Jillian Harvatt', 'Engineer I', 'jharvatt3@ibm.com', '0473232075', 'www.jharvatt3.com.au', 1),
-(5, 'rlzd1Hh', '2020-06-06 06:59:19', 1, 'Bednar-Brekke', '048 Green Lane', 'Silvan Blanpein', 'Nuclear Power Engineer', 'sblanpein4@github.com', '0445975651', 'www.sblanpein4.com.au', 1);